Add new domain custom resources
